Added Jupyter Notebook for GPIO controls
After a thorough search for information on how to easily control a MIO GPIO control from a Jupyter Notebook, I found a there was a general lack of information on such a seemingly simple topic.  This was the result of some experiments done on my Ultra96 to control an external circuit using a GPIO from the 40-pin expansion header.